factorized add timer form + integration of timer list in event tab
