+++ boardgame-ui/src/locales/translations.csv
@@ -174,7 +174,7 @@ timer.type.count_down,Count down,
 timer.type.reload,Reload,
 timer.your_turn,It is your turn!,
 timer.add-edit.title,Edit timer,
-timer.add-edit.timer.title,Timer,
+timer.add-edit.timer.title,Create timer,
 timer.add-edit.timer.duration,Initial timer value (ms),
 timer.add-edit.timer.type,Timer type,
 timer.add-edit.timer.reload-increment,Time increment,
