+++ boardgame-ui/src/pages/TimersPage.vue
@@ -1,7 +1,18 @@
 <template>
   <div>
     <HeroTitlePageLayout :title="$t('timers.title')"/>
-    <timer-list :timers="this.timers" />
+    <div class="container">
+      <section class="section">
+        <div class="columns">
+          <div class="column has-text-right">
+            <router-link tag="button" class="button is-primary" :to="{name: 'createtimer'}">
+              {{$t("timers.add")}}
+            </router-link>
+          </div>
+        </div>
+        <timer-list :timers="this.timers" />
+      </section>
+    </div>
   </div>
 </template>
